Add Cohere ASR support through generic QAIC registration
